+++ backend/agentd/agentd.go
@@ -55,8 +55,6 @@ func (a *Agentd) Start() error {
 
 	a.errChan = make(chan error, 1)
 
-	// handler := http.HandlerFunc(a.webSocketHandler)
-
 	// TODO: add JWT authentication support
 	handler := middlewares.BasicAuthentication(http.HandlerFunc(a.webSocketHandler), a.Store)