+++ backend/messaging/wizard_bus.go
@@ -2,7 +2,6 @@ package messaging
 
 import (
 	"errors"
-	"strings"
 	"sync"
 	"sync/atomic"
 
@@ -149,8 +148,4 @@ func (b *WizardBus) Publish(topic string, msg interface{}) error {
 	}
 
 	return nil
-}
-
-func topicIsDirect(s string) bool {
-	return len(strings.Split(s, ":")) == 5
 }