+++ backend/store/etcd/ring.go
@@ -1,6 +1,8 @@
 package etcd
 
-import "github.com/sensu/sensu-go/backend/ring"
+import (
+	"github.com/sensu/sensu-go/backend/ring"
+)
 
 // GetRing gets a named Ring.
 func (s *Store) GetRing(path ...string) ring.Interface {